+++ x/dex/keeper/msg_server_test.go
@@ -32,19 +32,19 @@ func TestCreatePool(t *testing.T) {
 		{
 			name:        "invalid creator addr",
 			creatorAddr: []byte{},
-			poolParams:  types.PoolParams{},
+			poolParams:  types.PoolParams{PoolType: types.PoolType_BALANCER},
 			poolAssets:  []types.PoolAsset{},
 			expectedErr: fmt.Errorf("empty address string is not allowed"),
 		},
 		{
 			name:        "not enough assets",
-			poolParams:  types.PoolParams{},
+			poolParams:  types.PoolParams{PoolType: types.PoolType_BALANCER},
 			poolAssets:  []types.PoolAsset{},
 			expectedErr: types.ErrTooFewPoolAssets,
 		},
 		{
 			name:       "too many assets",
-			poolParams: types.PoolParams{},
+			poolParams: types.PoolParams{PoolType: types.PoolType_BALANCER},
 			poolAssets: []types.PoolAsset{
 				{
 					Token:  sdk.NewInt64Coin("aaa", 1),
@@ -63,7 +63,7 @@ func TestCreatePool(t *testing.T) {
 		},
 		{
 			name:       "asset not whitelisted 1",
-			poolParams: types.PoolParams{},
+			poolParams: types.PoolParams{PoolType: types.PoolType_BALANCER},
 			poolAssets: []types.PoolAsset{
 				{
 					Token:  sdk.NewInt64Coin(common.DenomUSDC, 1),
@@ -78,7 +78,7 @@ func TestCreatePool(t *testing.T) {
 		},
 		{
 			name:       "asset not whitelisted 2",
-			poolParams: types.PoolParams{},
+			poolParams: types.PoolParams{PoolType: types.PoolType_BALANCER},
 			poolAssets: []types.PoolAsset{
 				{
 					Token:  sdk.NewInt64Coin("aaa", 1),
@@ -93,7 +93,7 @@ func TestCreatePool(t *testing.T) {
 		},
 		{
 			name:       "insufficient pool creation fee",
-			poolParams: types.PoolParams{},
+			poolParams: types.PoolParams{PoolType: types.PoolType_BALANCER},
 			poolAssets: []types.PoolAsset{
 				{
 					Token:  sdk.NewInt64Coin(common.DenomUSDC, 1),
@@ -113,7 +113,7 @@ func TestCreatePool(t *testing.T) {
 		},
 		{
 			name:       "insufficient initial deposit",
-			poolParams: types.PoolParams{},
+			poolParams: types.PoolParams{PoolType: types.PoolType_BALANCER},
 			poolAssets: []types.PoolAsset{
 				{
 					Token:  sdk.NewInt64Coin(common.DenomNIBI, 1),
@@ -131,7 +131,127 @@ func TestCreatePool(t *testing.T) {
 		},
 		{
 			name:       "successful pool creation",
-			poolParams: types.PoolParams{},
+			poolParams: types.PoolParams{PoolType: types.PoolType_BALANCER},
+			poolAssets: []types.PoolAsset{
+				{
+					Token:  sdk.NewInt64Coin(common.DenomNUSD, 1),
+					Weight: sdk.OneInt(),
+				},
+				{
+					Token:  sdk.NewInt64Coin(common.DenomUSDC, 1),
+					Weight: sdk.OneInt(),
+				},
+			},
+			senderInitialFunds: sdk.NewCoins(
+				sdk.NewInt64Coin(common.DenomNIBI, 1e9),
+				sdk.NewInt64Coin(common.DenomNUSD, 1),
+				sdk.NewInt64Coin(common.DenomUSDC, 1),
+			),
+			expectedErr: nil,
+		},
+		{
+			name:        "invalid creator addr - Stableswap",
+			creatorAddr: []byte{},
+			poolParams:  types.PoolParams{PoolType: types.PoolType_STABLESWAP, A: sdk.OneInt()},
+			poolAssets:  []types.PoolAsset{},
+			expectedErr: fmt.Errorf("empty address string is not allowed"),
+		},
+		{
+			name:        "not enough assets - Stableswap",
+			poolParams:  types.PoolParams{PoolType: types.PoolType_STABLESWAP, A: sdk.OneInt()},
+			poolAssets:  []types.PoolAsset{},
+			expectedErr: types.ErrTooFewPoolAssets,
+		},
+		{
+			name:       "too many assets - Stableswap",
+			poolParams: types.PoolParams{PoolType: types.PoolType_STABLESWAP, A: sdk.OneInt()},
+			poolAssets: []types.PoolAsset{
+				{
+					Token:  sdk.NewInt64Coin("aaa", 1),
+					Weight: sdk.OneInt(),
+				},
+				{
+					Token:  sdk.NewInt64Coin("bbb", 1),
+					Weight: sdk.OneInt(),
+				},
+				{
+					Token:  sdk.NewInt64Coin("ccc", 1),
+					Weight: sdk.OneInt(),
+				},
+			},
+			expectedErr: types.ErrTooManyPoolAssets,
+		},
+		{
+			name:       "asset not whitelisted 1 - Stableswap",
+			poolParams: types.PoolParams{PoolType: types.PoolType_STABLESWAP, A: sdk.OneInt()},
+			poolAssets: []types.PoolAsset{
+				{
+					Token:  sdk.NewInt64Coin(common.DenomUSDC, 1),
+					Weight: sdk.OneInt(),
+				},
+				{
+					Token:  sdk.NewInt64Coin("aaaa", 1),
+					Weight: sdk.OneInt(),
+				},
+			},
+			expectedErr: types.ErrTokenNotAllowed,
+		},
+		{
+			name:       "asset not whitelisted 2 - Stableswap",
+			poolParams: types.PoolParams{PoolType: types.PoolType_STABLESWAP, A: sdk.OneInt()},
+			poolAssets: []types.PoolAsset{
+				{
+					Token:  sdk.NewInt64Coin("aaa", 1),
+					Weight: sdk.OneInt(),
+				},
+				{
+					Token:  sdk.NewInt64Coin(common.DenomUSDC, 1),
+					Weight: sdk.OneInt(),
+				},
+			},
+			expectedErr: types.ErrTokenNotAllowed,
+		},
+		{
+			name:       "insufficient pool creation fee - Stableswap",
+			poolParams: types.PoolParams{PoolType: types.PoolType_STABLESWAP, A: sdk.OneInt()},
+			poolAssets: []types.PoolAsset{
+				{
+					Token:  sdk.NewInt64Coin(common.DenomUSDC, 1),
+					Weight: sdk.OneInt(),
+				},
+				{
+					Token:  sdk.NewInt64Coin(common.DenomNUSD, 1),
+					Weight: sdk.OneInt(),
+				},
+			},
+			senderInitialFunds: sdk.NewCoins(
+				sdk.NewInt64Coin("unibi", 1e9-1),
+				sdk.NewInt64Coin("aaa", 1),
+				sdk.NewInt64Coin("bbb", 1),
+			),
+			expectedErr: fmt.Errorf("999999999unibi is smaller than 1000000000unibi: insufficient funds"),
+		},
+		{
+			name:       "insufficient initial deposit - Stableswap",
+			poolParams: types.PoolParams{PoolType: types.PoolType_STABLESWAP, A: sdk.OneInt()},
+			poolAssets: []types.PoolAsset{
+				{
+					Token:  sdk.NewInt64Coin(common.DenomNIBI, 1),
+					Weight: sdk.OneInt(),
+				},
+				{
+					Token:  sdk.NewInt64Coin(common.DenomUSDC, 1),
+					Weight: sdk.OneInt(),
+				},
+			},
+			senderInitialFunds: sdk.NewCoins(
+				sdk.NewInt64Coin("unibi", 1e9),
+			),
+			expectedErr: fmt.Errorf("0unibi is smaller than 1unibi: insufficient funds"),
+		},
+		{
+			name:       "successful pool creation - Stableswap",
+			poolParams: types.PoolParams{PoolType: types.PoolType_STABLESWAP, A: sdk.OneInt()},
 			poolAssets: []types.PoolAsset{
 				{
 					Token:  sdk.NewInt64Coin(common.DenomNUSD, 1),
@@ -294,6 +414,98 @@ func TestMsgServerJoinPool(t *testing.T) {
 				),
 				/*shares=*/ 150),
 		},
+		{
+			name: "join with all assets - Stablepool",
+			joinerInitialFunds: sdk.NewCoins(
+				sdk.NewInt64Coin("unibi", 100),
+				sdk.NewInt64Coin(common.DenomNUSD, 100),
+			),
+			initialPool: mock.DexStablePool(
+				/*poolId=*/ 1,
+				/*assets=*/ sdk.NewCoins(
+					sdk.NewInt64Coin("unibi", 100),
+					sdk.NewInt64Coin(common.DenomNUSD, 100),
+				),
+				/*shares=*/ 100),
+			tokensIn: sdk.NewCoins(
+				sdk.NewInt64Coin("unibi", 100),
+				sdk.NewInt64Coin(common.DenomNUSD, 100),
+			),
+			expectedNumSharesOut:     sdk.NewInt64Coin(shareDenom, 100),
+			expectedRemCoins:         sdk.NewCoins(),
+			expectedJoinerFinalFunds: sdk.NewCoins(sdk.NewInt64Coin(shareDenom, 100)),
+			expectedFinalPool: mock.DexStablePool(
+				/*poolId=*/ 1,
+				/*assets=*/ sdk.NewCoins(
+					sdk.NewInt64Coin("unibi", 200),
+					sdk.NewInt64Coin(common.DenomNUSD, 200),
+				),
+				/*shares=*/ 200),
+		},
+		{
+			name: "join with some assets, none remaining - Stablepool",
+			joinerInitialFunds: sdk.NewCoins(
+				sdk.NewInt64Coin("unibi", 100),
+				sdk.NewInt64Coin(common.DenomNUSD, 100),
+			),
+			initialPool: mock.DexStablePool(
+				/*poolId=*/ 1,
+				/*assets=*/ sdk.NewCoins(
+					sdk.NewInt64Coin("unibi", 100),
+					sdk.NewInt64Coin(common.DenomNUSD, 100),
+				),
+				/*shares=*/ 100),
+			tokensIn: sdk.NewCoins(
+				sdk.NewInt64Coin("unibi", 50),
+				sdk.NewInt64Coin(common.DenomNUSD, 50),
+			),
+			expectedNumSharesOut: sdk.NewInt64Coin(shareDenom, 50),
+			expectedRemCoins:     []sdk.Coin{},
+			expectedJoinerFinalFunds: sdk.NewCoins(
+				sdk.NewInt64Coin(shareDenom, 50),
+				sdk.NewInt64Coin("unibi", 50),
+				sdk.NewInt64Coin(common.DenomNUSD, 50),
+			),
+			expectedFinalPool: mock.DexStablePool(
+				/*poolId=*/ 1,
+				/*assets=*/ sdk.NewCoins(
+					sdk.NewInt64Coin("unibi", 150),
+					sdk.NewInt64Coin(common.DenomNUSD, 150),
+				),
+				/*shares=*/ 150),
+		},
+		{
+			name: "join with some assets, some remaining - Stablepool",
+			joinerInitialFunds: sdk.NewCoins(
+				sdk.NewInt64Coin("unibi", 100),
+				sdk.NewInt64Coin(common.DenomNUSD, 100),
+			),
+			initialPool: mock.DexStablePool(
+				/*poolId=*/ 1,
+				/*assets=*/ sdk.NewCoins(
+					sdk.NewInt64Coin("unibi", 100),
+					sdk.NewInt64Coin(common.DenomNUSD, 100),
+				),
+				/*shares=*/ 100),
+			tokensIn: sdk.NewCoins(
+				sdk.NewInt64Coin("unibi", 50),
+				sdk.NewInt64Coin(common.DenomNUSD, 75),
+			),
+			expectedNumSharesOut: sdk.NewInt64Coin(shareDenom, 62),
+			expectedRemCoins:     []sdk.Coin{},
+			expectedJoinerFinalFunds: sdk.NewCoins(
+				sdk.NewInt64Coin(shareDenom, 62),
+				sdk.NewInt64Coin("unibi", 50),
+				sdk.NewInt64Coin(common.DenomNUSD, 25),
+			),
+			expectedFinalPool: mock.DexStablePool(
+				/*poolId=*/ 1,
+				/*assets=*/ sdk.NewCoins(
+					sdk.NewInt64Coin("unibi", 150),
+					sdk.NewInt64Coin(common.DenomNUSD, 175),
+				),
+				/*shares=*/ 162),
+		},
 	}
 
 	for _, tc := range tests {
@@ -413,6 +625,81 @@ func TestMsgServerExitPool(t *testing.T) {
 				sdk.NewInt64Coin(shareDenom, 50),
 			),
 			expectedFinalPool: mock.DexPool(
+				/*poolId=*/ 1,
+				/*assets=*/ sdk.NewCoins(
+					sdk.NewInt64Coin("unibi", 51),
+					sdk.NewInt64Coin(common.DenomNUSD, 51),
+				),
+				/*shares=*/ 50,
+			),
+		},
+		{
+			name: "exit all pool shares - StablePool",
+			joinerInitialFunds: sdk.NewCoins(
+				sdk.NewInt64Coin("unibi", 100),
+				sdk.NewInt64Coin(common.DenomNUSD, 100),
+				sdk.NewInt64Coin(shareDenom, 100),
+			),
+			initialPoolFunds: sdk.NewCoins(
+				sdk.NewInt64Coin("unibi", 100),
+				sdk.NewInt64Coin(common.DenomNUSD, 100),
+			),
+			initialPool: mock.DexStablePool(
+				/*poolId=*/ 1,
+				/*assets=*/ sdk.NewCoins(
+					sdk.NewInt64Coin("unibi", 100),
+					sdk.NewInt64Coin(common.DenomNUSD, 100),
+				),
+				/*shares=*/ 100,
+			),
+			poolSharesIn: sdk.NewInt64Coin(shareDenom, 100),
+			expectedTokensOut: sdk.NewCoins(
+				sdk.NewInt64Coin("unibi", 99),
+				sdk.NewInt64Coin(common.DenomNUSD, 99),
+			),
+			expectedJoinerFinalFunds: sdk.NewCoins(
+				sdk.NewInt64Coin("unibi", 199),
+				sdk.NewInt64Coin(common.DenomNUSD, 199),
+			),
+			expectedFinalPool: mock.DexStablePool(
+				/*poolId=*/ 1,
+				/*assets=*/ sdk.NewCoins(
+					sdk.NewInt64Coin("unibi", 1),
+					sdk.NewInt64Coin(common.DenomNUSD, 1),
+				),
+				/*shares=*/ 0,
+			),
+		},
+		{
+			name: "exit half pool shares - StablePool",
+			joinerInitialFunds: sdk.NewCoins(
+				sdk.NewInt64Coin("unibi", 100),
+				sdk.NewInt64Coin(common.DenomNUSD, 100),
+				sdk.NewInt64Coin(shareDenom, 100),
+			),
+			initialPoolFunds: sdk.NewCoins(
+				sdk.NewInt64Coin("unibi", 100),
+				sdk.NewInt64Coin(common.DenomNUSD, 100),
+			),
+			initialPool: mock.DexStablePool(
+				/*poolId=*/ 1,
+				/*assets=*/ sdk.NewCoins(
+					sdk.NewInt64Coin("unibi", 100),
+					sdk.NewInt64Coin(common.DenomNUSD, 100),
+				),
+				/*shares=*/ 100,
+			),
+			poolSharesIn: sdk.NewInt64Coin(shareDenom, 50),
+			expectedTokensOut: sdk.NewCoins(
+				sdk.NewInt64Coin("unibi", 49),
+				sdk.NewInt64Coin(common.DenomNUSD, 49),
+			),
+			expectedJoinerFinalFunds: sdk.NewCoins(
+				sdk.NewInt64Coin("unibi", 149),
+				sdk.NewInt64Coin(common.DenomNUSD, 149),
+				sdk.NewInt64Coin(shareDenom, 50),
+			),
+			expectedFinalPool: mock.DexStablePool(
 				/*poolId=*/ 1,
 				/*assets=*/ sdk.NewCoins(
 					sdk.NewInt64Coin("unibi", 51),
@@ -614,6 +901,147 @@ func TestMsgServerSwapAssets(t *testing.T) {
 				sdk.NewInt64Coin("unibi", 100),
 			),
 			expectedFinalPool: mock.DexPool(
+				/*poolId=*/ 1,
+				/*assets=*/ sdk.NewCoins(
+					sdk.NewInt64Coin("unibi", 100),
+					sdk.NewInt64Coin(common.DenomNUSD, 100),
+				),
+				/*shares=*/ 100,
+			),
+			expectedError: types.ErrSameTokenDenom,
+		},
+		{
+			name: "regular swap - StableSwap",
+			userInitialFunds: sdk.NewCoins(
+				sdk.NewInt64Coin("unibi", 100),
+			),
+			initialPool: mock.DexStablePool(
+				/*poolId=*/ 1,
+				/*assets=*/ sdk.NewCoins(
+					sdk.NewInt64Coin("unibi", 100),
+					sdk.NewInt64Coin(common.DenomNUSD, 100),
+				),
+				/*shares=*/ 100,
+			),
+			tokenIn:          sdk.NewInt64Coin("unibi", 100),
+			tokenOutDenom:    common.DenomNUSD,
+			expectedTokenOut: sdk.NewInt64Coin(common.DenomNUSD, 30),
+			expectedUserFinalFunds: sdk.NewCoins(
+				sdk.NewInt64Coin(common.DenomNUSD, 30),
+			),
+			expectedFinalPool: mock.DexStablePool(
+				/*poolId=*/ 1,
+				/*assets=*/ sdk.NewCoins(
+					sdk.NewInt64Coin("unibi", 200),
+					sdk.NewInt64Coin(common.DenomNUSD, 70),
+				),
+				/*shares=*/ 100,
+			),
+			expectedError: nil,
+		},
+		{
+			name: "not enough user funds - StableSwap",
+			userInitialFunds: sdk.NewCoins(
+				sdk.NewInt64Coin("unibi", 1),
+			),
+			initialPool: mock.DexStablePool(
+				/*poolId=*/ 1,
+				/*assets=*/ sdk.NewCoins(
+					sdk.NewInt64Coin("unibi", 100),
+					sdk.NewInt64Coin(common.DenomNUSD, 100),
+				),
+				/*shares=*/ 100,
+			),
+			tokenIn:       sdk.NewInt64Coin("unibi", 100),
+			tokenOutDenom: common.DenomNUSD,
+			expectedUserFinalFunds: sdk.NewCoins(
+				sdk.NewInt64Coin("unibi", 1),
+			),
+			expectedFinalPool: mock.DexStablePool(
+				/*poolId=*/ 1,
+				/*assets=*/ sdk.NewCoins(
+					sdk.NewInt64Coin("unibi", 100),
+					sdk.NewInt64Coin(common.DenomNUSD, 100),
+				),
+				/*shares=*/ 100,
+			),
+			expectedError: sdkerrors.ErrInsufficientFunds,
+		},
+		{
+			name: "invalid token in denom - StableSwap",
+			userInitialFunds: sdk.NewCoins(
+				sdk.NewInt64Coin("foo", 100),
+			),
+			initialPool: mock.DexStablePool(
+				/*poolId=*/ 1,
+				/*assets=*/ sdk.NewCoins(
+					sdk.NewInt64Coin("unibi", 100),
+					sdk.NewInt64Coin(common.DenomNUSD, 100),
+				),
+				/*shares=*/ 100,
+			),
+			tokenIn:       sdk.NewInt64Coin("foo", 100),
+			tokenOutDenom: common.DenomNUSD,
+			expectedUserFinalFunds: sdk.NewCoins(
+				sdk.NewInt64Coin("foo", 100),
+			),
+			expectedFinalPool: mock.DexStablePool(
+				/*poolId=*/ 1,
+				/*assets=*/ sdk.NewCoins(
+					sdk.NewInt64Coin("unibi", 100),
+					sdk.NewInt64Coin(common.DenomNUSD, 100),
+				),
+				/*shares=*/ 100,
+			),
+			expectedError: types.ErrTokenDenomNotFound,
+		},
+		{
+			name: "invalid token out denom - StableSwap",
+			userInitialFunds: sdk.NewCoins(
+				sdk.NewInt64Coin("unibi", 100),
+			),
+			initialPool: mock.DexStablePool(
+				/*poolId=*/ 1,
+				/*assets=*/ sdk.NewCoins(
+					sdk.NewInt64Coin("unibi", 100),
+					sdk.NewInt64Coin(common.DenomNUSD, 100),
+				),
+				/*shares=*/ 100,
+			),
+			tokenIn:       sdk.NewInt64Coin("unibi", 100),
+			tokenOutDenom: "foo",
+			expectedUserFinalFunds: sdk.NewCoins(
+				sdk.NewInt64Coin("unibi", 100),
+			),
+			expectedFinalPool: mock.DexStablePool(
+				/*poolId=*/ 1,
+				/*assets=*/ sdk.NewCoins(
+					sdk.NewInt64Coin("unibi", 100),
+					sdk.NewInt64Coin(common.DenomNUSD, 100),
+				),
+				/*shares=*/ 100,
+			),
+			expectedError: types.ErrTokenDenomNotFound,
+		},
+		{
+			name: "same token in and token out denom - StableSwap",
+			userInitialFunds: sdk.NewCoins(
+				sdk.NewInt64Coin("unibi", 100),
+			),
+			initialPool: mock.DexStablePool(
+				/*poolId=*/ 1,
+				/*assets=*/ sdk.NewCoins(
+					sdk.NewInt64Coin("unibi", 100),
+					sdk.NewInt64Coin(common.DenomNUSD, 100),
+				),
+				/*shares=*/ 100,
+			),
+			tokenIn:       sdk.NewInt64Coin("unibi", 100),
+			tokenOutDenom: "unibi",
+			expectedUserFinalFunds: sdk.NewCoins(
+				sdk.NewInt64Coin("unibi", 100),
+			),
+			expectedFinalPool: mock.DexStablePool(
 				/*poolId=*/ 1,
 				/*assets=*/ sdk.NewCoins(
 					sdk.NewInt64Coin("unibi", 100),