+++ x/dex/simulation/operations.go
@@ -294,8 +294,9 @@ func genBalancerPoolParams(r *rand.Rand, blockTime time.Time, assets []types.Poo
 	// TODO: Randomly generate LBP params
 	return types.PoolParams{
 		// SwapFee:                  swapFee,
-		SwapFee: sdk.ZeroDec(),
-		ExitFee: exitFee,
+		SwapFee:  sdk.ZeroDec(),
+		ExitFee:  exitFee,
+		PoolType: types.PoolType_BALANCER,
 	}
 }